+++ packages/data-types/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/interfaces/base-crop.interface';
+export * from './lib/interfaces/consumable.interface';
 export * from './lib/interfaces/cooking-recipe.interface';
 export * from './lib/interfaces/crafting-recipe.interface';
 export * from './lib/interfaces/critter.interface';
@@ -14,8 +15,13 @@ export * from './lib/interfaces/journal-order.interface';
 export * from './lib/interfaces/npc.interface';
 export * from './lib/interfaces/tag-based-item.interface';
 export * from './lib/interfaces/unlock-by-mastery.interface';
+export * from './lib/interfaces/offering-altar.interface';
+export * from './lib/interfaces/offerings.interface';
+export * from './lib/interfaces/offering-reward.interface';
 
 export * from './lib/enums/quality.enum';
+export * from './lib/enums/season.enum';
+export * from './lib/enums/weather.enum';
 
 export * from './lib/types/minimal-item.type';
 export * from './lib/types/gift-preference.type';