+++ packages/data-types/src/lib/interfaces/offering-altar.interface.ts
@@ -1,0 +1,8 @@
+import { Offerings } from "./offerings.interface";
+
+export interface OfferingAltar {
+    key: string;
+    offeringGroupTitle: string;
+    offeringGroupRewardText: string;
+    offerings: Offerings[]
+}