+++ packages/guide/src/app/crafting/components/tables/inventory-table/inventory-table.component.ts
@@ -5,15 +5,25 @@ import { CraftingRecipe } from "@ci/data-types";
 @Component({
     selector: 'app-inventory-table',
     templateUrl: './inventory-table.component.html',
-    styleUrls: ['./inventory-table.component.scss'],
 })
 export class InventoryTableComponent extends BaseTableComponent<CraftingRecipe> {
     protected readonly BASE_DISPLAY_COLUMNS: string[] = [
         'icon',
         'outputName',
         'ingredients',
-        'price',
+        'sellPrice',
         'unlock'
     ];
+
+
+    override sortingDataAccessor = (item: CraftingRecipe, property: string) => {
+
+        const sortHelperValue = this.sortHelper(item.item, property)
+
+        if (sortHelperValue !== null) return sortHelperValue;
+
+        return 0;
+
+    };
 
 }