+++ packages/guide/src/app/database/components/database-crafting/database-crafting.component.ts
@@ -5,7 +5,6 @@ import { CraftingRecipe, Item } from "@ci/data-types";
 @Component({
     selector: 'app-database-crafting',
     templateUrl: './database-crafting.component.html',
-    styleUrls: ['./database-crafting.component.scss'],
     encapsulation: ViewEncapsulation.None
 })
 export class DatabaseCraftingComponent extends BaseDatabaseDetailPartComponent implements OnInit {