+++ packages/guide/src/app/shared/enums/ui-icon.enum.ts
@@ -9,6 +9,7 @@ export enum UiIcon {
     CRAFTING = 'T_Icon_Crafting',
     PEOPLE = 'T_Icon_People',
     DATABASE = 'T_Icon_AllQuest',
+    MAP = 'T_Icon_Map',
     TO_DO = 'T_Icon_Errands',
     BURGER_MENU = 'T_Icon_BurgerMenu',
 
@@ -17,6 +18,7 @@ export enum UiIcon {
     FARMING = 'T_Icon_FarmingM',
     ARTISAN = 'T_Icon_ArtisanM',
     COOKING = 'T_Icon_ConsumablesM',
+    OFFERING = 'T_Icon_Offering',
 
     //
     GIFTING = 'Icons_Gift',