+++ packages/guide/src/app/shared/pipes/is-item.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { IsItemPipe } from './is-item.pipe';
+
+describe('IsItemPipe', () => {
+    it('create an instance', () => {
+        const pipe = new IsItemPipe();
+        expect(pipe).toBeTruthy();
+    });
+});