+++ packages/pak-files-parser/src/interfaces/offering-match.interface.ts
@@ -1,0 +1,9 @@
+import { OfferingReward } from "@ci/data-types";
+
+export interface OfferingMatch {
+    key: string;
+    offeringId: string
+    rewardID: string
+    rewardItem: string
+    rewards: OfferingReward
+}