+++ packages/pak-files-parser/src/interfaces/raw-offering-reward.interface.ts
@@ -1,0 +1,26 @@
+export interface RawOfferingReward {
+    "offeringId": {
+        "DataTable": {
+            "ObjectName": string
+            "ObjectPath": string
+        },
+        "RowName": string
+    },
+    "rewardID": string
+    "description": {
+        "CultureInvariantString": null
+    },
+    "useItemMesh": boolean,
+    "scaleItemMesh": boolean,
+    "rewardItem": {
+        "data": {
+            "DataTable": {
+                "ObjectName": string
+                "ObjectPath": string
+            },
+            "RowName": string
+        },
+        "itemID": string;
+    },
+    "rewardMesh": null
+}