+++ packages/pak-files-parser/src/main.ts
@@ -20,6 +20,8 @@ import { CookingRecipeUnlockedByMasteryDbGenerator } from "./app/cooking-recipe-
 import { CookingDbGenerator } from "./app/cooking-db.generator";
 import { FruitTreeDbGenerator } from "./app/fruit-tree-db.generator";
 import { FruitPlantDbGenerator } from "./app/fruit-plant-db.generator";
+import { OfferingsDbGenerator } from "./app/offerings-db.generator";
+import { ConsumablesDbGenerator } from "./app/consumables-db.generator";
 
 const itemIconPath = config.itemIconPath
 const texturePath = config.texturePath;
@@ -39,6 +41,9 @@ const cookingRecipeUnlockedByMasteryDbMap = cookingRecipeUnlockedByMasteryDbGene
 const tagBasedItemsDbGenerator = new TagBasedItemGenericDbGenerator();
 const tagBasedItemsDbMap = tagBasedItemsDbGenerator.generate();
 
+const cookingDbGenerator = new CookingDbGenerator(itemDbMap, cookingRecipeUnlockedByMasteryDbMap, tagBasedItemsDbMap);
+const cookingDbMap = cookingDbGenerator.generate();
+
 const generators: Record<string, { generate: () => Map<string, any> }> = {
     'crafting-recipes': new CraftingRecipeDbGenerator(itemDbMap, craftingRecipeUnlockedByMasteryDbMap),
     'bugs-and-insects': new BugsAndInsectsDbGenerator(itemDbMap),
@@ -48,7 +53,6 @@ const generators: Record<string, { generate: () => Map<string, any> }> = {
     'fruit-trees': new FruitTreeDbGenerator(itemDbMap),
     'fruit-plants': new FruitPlantDbGenerator(itemDbMap),
     'item-processing': new ItemProcessorDbGenerator(itemDbMap),
-    'cooking-recipes': new CookingDbGenerator(itemDbMap, cookingRecipeUnlockedByMasteryDbMap, tagBasedItemsDbMap),
 
     'journal-fish': new JournalOrderDbGenerator('Caught/DT_JournalFish.json'),
     'journal-insects': new JournalOrderDbGenerator('Caught/DT_JournalInsects.json'),
@@ -64,13 +68,17 @@ const generators: Record<string, { generate: () => Map<string, any> }> = {
     'journal-artisan-products': new JournalOrderDbGenerator('Produce/DT_JournalArtisanProducts.json'),
     'journal-crops': new JournalOrderDbGenerator('Produce/DT_JournalCrops.json'),
 
+    'offerings': new OfferingsDbGenerator(itemDbMap, cookingDbMap),
+
     'gift-preferences': new GiftPreferencesDbGenerator(itemDbMap, npcDbMap),
     'npcs': new NPCDbGenerator(itemDbMap),
 
+    'consumables': new ConsumablesDbGenerator(),
 
     'tag-based-items': {generate: () => tagBasedItemsDbMap},
     'crafting-unlocks-by-mastery': {generate: () => craftingRecipeUnlockedByMasteryDbMap},
     'cooking-unlocks-by-mastery': {generate: () => cookingRecipeUnlockedByMasteryDbMap},
+    'cooking-recipes': {generate: () => cookingDbMap},
 
     // last so applied changed will be written as well
     items: {generate: () => itemDbMap},
@@ -147,7 +155,7 @@ async function extractImages() {
 
 
             counter++;
-            if (counter % 100 === 0) {
+            if (counter % 500 === 0) {
                 console.log(`processed ${counter} images...`);
             }
 
@@ -164,9 +172,18 @@ extractImages();
 Object.keys(generators).forEach(generatorName => {
         const generatedMap = generators[generatorName].generate();
 
-        // if(generatorName === 'fish')
-        //     printFish([...generatedMap.values()])
+        // if(generatorName === 'consumables')
+        // {
+        //     // @ts-ignore
+        //     console.log(generators['consumables'].buff);
+        //     // @ts-ignore
+        //     console.log(generators['consumables'].level);
+        //     // @ts-ignore
+        //     console.log(generators['consumables'].duration);
+        //     // @ts-ignore
+        //     console.log(generators['consumables'].itemType);
+        // }
 
-    generateJson(`${generatorName}.json`, [...generatedMap.values()], true);
+        generateJson(`${generatorName}.json`, [...generatedMap.values()], true);
     }
 );